% Auto-generated: do not edit by hand
\name{''Edit}

\alias{''Edit}

\title{Edit component}

\description{

}

\usage{
''Edit(id=NULL, note=NULL, dyn=NULL, scale=NULL,
accidental_margin=NULL, hide_toolbar=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{note}{Numeric. }

\item{dyn}{Character. }

\item{scale}{Numeric. }

\item{accidental_margin}{Numeric. }

\item{hide_toolbar}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

