% Auto-generated: do not edit by hand
\name{''Masking}

\alias{''Masking}

\title{Masking component}

\description{

}

\usage{
''Masking(id=NULL, masking_notevalues=NULL, masking_notesizes=NULL,
masking_colors=NULL, target_notevalues=NULL,
target_notesizes=NULL, target_colors=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{masking_notevalues}{Unnamed list. A label that will be printed when this component is rendered.}

\item{masking_notesizes}{Unnamed list. }

\item{masking_colors}{Unnamed list. }

\item{target_notevalues}{Unnamed list. }

\item{target_notesizes}{Unnamed list. }

\item{target_colors}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

