% Auto-generated: do not edit by hand
\name{''Orchestration functional no update}

\alias{''Orchestration functional no update}

\title{Orchestration functional no update component}

\description{

}

\usage{
''Orchestration functional no update(id=NULL, notes=NULL, target=NULL, instruments=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{notes}{Unnamed list. A label that will be printed when this component is rendered.}

\item{target}{Unnamed list. }

\item{instruments}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

