% Auto-generated: do not edit by hand
\name{''Orchestration}

\alias{''Orchestration}

\title{Orchestration component}

\description{

}

\usage{
''Orchestration(id=NULL, notes=NULL, target=NULL, instruments=NULL,
highlights=NULL, width=NULL, height=NULL, t_score_y=NULL,
b_score_y=NULL, target_color=NULL, text_space=NULL,
indexing=NULL, onClick=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{notes}{Unnamed list. A label that will be printed when this component is rendered.}

\item{target}{Unnamed list. }

\item{instruments}{Unnamed list. }

\item{highlights}{Unnamed list. }

\item{width}{Numeric. }

\item{height}{Numeric. }

\item{t_score_y}{Numeric. }

\item{b_score_y}{Numeric. }

\item{target_color}{Character. }

\item{text_space}{Numeric. }

\item{indexing}{Logical. }

\item{onClick}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

