% Auto-generated: do not edit by hand
\name{''Pianoinput}

\alias{''Pianoinput}

\title{Pianoinput component}

\description{

}

\usage{
''Pianoinput(id=NULL, note=NULL, scale=NULL, keywidth=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{note}{Numeric. }

\item{scale}{Numeric. }

\item{keywidth}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

