% Auto-generated: do not edit by hand
\name{''Pianoroll}

\alias{''Pianoroll}

\title{Pianoroll component}

\description{

}

\usage{
''Pianoroll(id=NULL, width=NULL, height=NULL, stave_spacing=NULL,
scale=NULL, bar_offset=NULL, onHover=NULL, onClick=NULL,
stave_list=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{width}{Numeric. A label that will be printed when this component is rendered.}

\item{height}{Numeric. }

\item{stave_spacing}{Numeric. }

\item{scale}{Numeric. }

\item{bar_offset}{Numeric. }

\item{onHover}{Named list. }

\item{onClick}{Named list. }

\item{stave_list}{Unnamed list. A Array containing object to hold bars as [[bar1],[bar2] etc..].
Bars containing staves as {clef:'string', name:'string', 'notes':[array or nested array of notenames], colors:[array that matches notes-array]}}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

