% Auto-generated: do not edit by hand
\name{''Range - working Copy without colors}

\alias{''Range - working Copy without colors}

\title{Range - working Copy without colors component}

\description{

}

\usage{
''Range - working Copy without colors(id=NULL, width=NULL, height=NULL, notes=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{width}{Numeric. A label that will be printed when this component is rendered.}

\item{height}{Numeric. }

\item{notes}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

