% Auto-generated: do not edit by hand
\name{''Range}

\alias{''Range}

\title{Range component}

\description{

}

\usage{
''Range(id=NULL, width=NULL, height=NULL, staffY=NULL, notes=NULL,
clefs=NULL, colors=NULL, annotations=NULL, draw_lines=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{width}{Numeric. }

\item{height}{Numeric. }

\item{staffY}{Numeric. }

\item{notes}{Unnamed list. }

\item{clefs}{Unnamed list. }

\item{colors}{Unnamed list. }

\item{annotations}{Unnamed list. }

\item{draw_lines}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

