% Auto-generated: do not edit by hand
\name{''Stave}

\alias{''Stave}

\title{Stave component}

\description{

}

\usage{
''Stave(id=NULL, notes=NULL, clef=NULL, width=NULL, height=NULL,
scale=NULL, padding=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{notes}{Character. A label that will be printed when this component is rendered.}

\item{clef}{Character. }

\item{width}{Numeric. }

\item{height}{Numeric. }

\item{scale}{Numeric. }

\item{padding}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

